mptcp: remove duplication between select_size and current_mss
The idea behind the first part of mptcp_select_size and mptcp_current_mss is
the same except that the former uses mss_cache to compute the mss while the
latter use tcp_current_mss. This patch refactor the code such that it defines
a new function that take a callback function which returns the expected mss
value for a subflow.

Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>